+++ packages/demo/src/app/bootstrap/components/badge/badge-interactive-demo/badge-interactive-demo.component.ts
@@ -3,7 +3,6 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'app-badge-interactive-demo',
   templateUrl: './badge-interactive-demo.component.html',
-  styleUrls: ['./badge-interactive-demo.component.scss']
+  styleUrls: ['./badge-interactive-demo.component.scss'],
 })
-export class BadgeInteractiveDemoComponent {
-}
+export class BadgeInteractiveDemoComponent {}