+++ packages/demo/src/app/ng-bootstrap/components/accordion/accordion-demo/accordion-demo.component.ts
@@ -2,8 +2,6 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-ngb-accordion-demo',
-  templateUrl: './accordion-demo.component.html'
+  templateUrl: './accordion-demo.component.html',
 })
-export class NgbAccordionDemoComponent {
-
-}
+export class NgbAccordionDemoComponent {}