+++ packages/demo/src/app/post-sample/components/custom-select/custom-select-floating-demo/custom-select-floating-demo.component.ts
@@ -1,9 +1,9 @@
-import {Component, OnInit, Input} from '@angular/core';
-import {CustomSelectDemoComponent} from '../custom-select-demo/custom-select-demo.component';
+import { Component, OnInit, Input } from '@angular/core';
+import { CustomSelectDemoComponent } from '../custom-select-demo/custom-select-demo.component';
 
 @Component({
   selector: 'app-custom-select-floating-demo',
-  templateUrl: './custom-select-floating-demo.component.html'
+  templateUrl: './custom-select-floating-demo.component.html',
 })
 export class CustomSelectFloatingDemoComponent extends CustomSelectDemoComponent implements OnInit {
   @Input() public noSelected: boolean = false;