+++ packages/demo/src/app/post-sample/components/sizing/sizing-demo/sizing-demo.component.ts
@@ -2,18 +2,18 @@
  * Copyright 2021 by Swiss Post, Information Technology
  */
 
-import {Component, Input} from '@angular/core';
-import {UntypedFormBuilder, UntypedFormGroup} from '@angular/forms';
+import { Component, Input } from '@angular/core';
+import { UntypedFormBuilder, UntypedFormGroup } from '@angular/forms';
 
 @Component({
   selector: 'app-sizing-demo',
-  templateUrl: 'sizing-demo.component.html'
+  templateUrl: 'sizing-demo.component.html',
 })
 export class SizingDemoComponent {
   @Input() sizes: unknown;
 
   demoForm: UntypedFormGroup;
-  bootstrapSizes: {value: string, name: string}[];
+  bootstrapSizes: { value: string; name: string }[];
 
   constructor(private fb: UntypedFormBuilder) {
     this.demoForm = fb.group({
@@ -32,15 +32,15 @@ export class SizingDemoComponent {
       'max-height': fb.group({
         prefix: ['mh-'],
         size: ['100'],
-      })
+      }),
     });
 
     this.bootstrapSizes = [
-      {value: '25', name: '25%'},
-      {value: '50', name: '50%'},
-      {value: '75', name: '75%'},
-      {value: '100', name: '100%'},
-      {value: 'auto', name: 'Auto (default)'},
+      { value: '25', name: '25%' },
+      { value: '50', name: '50%' },
+      { value: '75', name: '75%' },
+      { value: '100', name: '100%' },
+      { value: 'auto', name: 'Auto (default)' },
     ];
   }
 
@@ -49,7 +49,9 @@ export class SizingDemoComponent {
   }
 
   get demoClass(): string {
-    return Object.keys(this.demoForm.value).map((prop) => this.getPrefix(prop) + this.getSize(prop)).join(' ');
+    return Object.keys(this.demoForm.value)
+      .map(prop => this.getPrefix(prop) + this.getSize(prop))
+      .join(' ');
   }
 
   getPrefix(prop: string): string {